refactor: <BalanceAndActivity />
