--- .yarnclean
@@ -1,1 +1,0 @@
-@types/react-native
