+++ package.json
@@ -97,7 +97,6 @@
     "react-router-dom": "6.0.0-beta.0",
     "safe-compare": "1.1.4",
     "schema-inspector": "2.0.1",
-    "swr": "0.5.6",
     "ts-debounce": "3.0.0",
     "use-events": "1.4.2",
     "use-latest": "1.2.0",
