+++ src/app.tsx
@@ -46,5 +46,3 @@ export const App: React.FC = () => {
     </ThemeProvider>
   );
 };
-
-export default App;
