+++ src/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './app';
+import { App } from './app';
 
 import { persistAndRenderApp } from '@common/persistence';
 import { initSentry } from '@common/sentry-init';
